fix: hiclass test failing
